Fixed the argument type of limit_virtual_memory(): should be unsigned long.

+++ demos/ppl_lpsol/ppl_lpsol.c
@@ -521,7 +521,7 @@ set_alarm_on_cpu_time(unsigned seconds, void (*handler)(int)) {
 #if PPL_HAVE_DECL_RLIMIT_AS
 
 void
-limit_virtual_memory(unsigned bytes) {
+limit_virtual_memory(unsigned long bytes) {
   struct rlimit t;
 
   if (getrlimit(RLIMIT_AS, &t) != 0)
@@ -537,7 +537,7 @@ limit_virtual_memory(unsigned bytes) {
 #else
 
 void
-limit_virtual_memory(unsigned bytes ATTRIBUTE_UNUSED) {
+limit_virtual_memory(unsigned long bytes ATTRIBUTE_UNUSED) {
 }
 
 #endif /* !PPL_HAVE_DECL_RLIMIT_AS */
